// src/api/model/messageDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * PulseChat.API
 * PulseChat REST API
 * OpenAPI spec version: v1
 */
import type { MessageType } from './messageType';

export interface MessageDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  roomId?: string | null;
  senderId?: string;
  /** @nullable */
  senderUsername?: string | null;
  /** @nullable */
  clientMessageId?: string | null;
  type?: MessageType;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  createdAt?: string;
}